pyproject.toml
[tool.poetry]
name = "health-chatbox"
version = "0.1.0"
description = ""
authors = ["Rom1009 <quynguyencurit2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
requests = "^2.32.3"
crewai = "^0.86.0"
streamlit = "^1.41.0"
fastapi = "^0.115.6"
bs4 = "^0.0.2"
uvicorn = "^0.22.0"
numpy = "^1.25.0" 
pandas = "^2.0.0" 
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"